--- src/main/java/com/app/smartdrive/entities/customer/Customer.java
@@ -1,5 +1,0 @@
-package com.app.smartdrive.entities.customer;
-
-public class Customer {
-    private String name;
-}
